Extract trace store id path helpers
